pyproject.toml
[project]
name = "djangocmsjoy"
version = "0.1.0"
description = "Django CMS Joy Development Project"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "django>=5.2,<5.3",
    "django-cms>=5.0,<5.1",
    "djangocms-admin-style>=3.3",
    "requests>=2.32.0",
    "djangorestframework>=3.15.0",
    "gunicorn>=23.0,<24",
    "django-bootstrap5>=25.0,<26",
    "django-allauth[socialaccount]>=65.0,<66",
    "psycopg2-binary>=2.9",
]

[dependency-groups]
dev = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["djangocmsjoy"]
